Fix nodeToString() to return valid HTML
In some cases the output of nodeToString() is not valid HTML, e.g. for iframe nodes.

Solution found here: http://stackoverflow.com/questions/27020950/lxml-modify-tags-prevent

An example of bad behavior is an article with a youtube video inside. The video embed code is transformed into `<iframe ... />`, that is not valid, and thus the result in the browser is a page with no visible content after the video.