Improved title extraction via h1 and og:title
This commit improves title extraction by relying also on h1 and og:title. 
The current extraction was flawed in multiple ways: wrong part extracted during splitting, truncated title or typos or wrong spellings (probably caused by a bad CMS used by the news source).

In short, combining title, og:title, and h1 allows reducing these errors. The general assumption is that h1 is the most polished/curated title, as it's visible to the users, but can be hard to extract (not always present, not sure if we're detecting the right h1). og:title often contains the title properly split, and is also a good "3rd party" to help choosing among title and h1.

Explicit rules are documented within the get_title() comments.

These rules have been created via empiric tests (over 60 sources, about 10 articles/source). It's easy to find/create counter examples, but in general the results are better than before :) (about 100 more correct results over 4-500 articles). I tried to find a minimum set of understandable/reasonable rules that could cover the majority of the tests I had.

Included in the patch, I'm removing the split by ':' because most of the time this removes relevant part of the title (e.g., 'Game of the Week: Random Name', always removes the most significant part, i.e. the game name, and similar for recipes, events...). This potentially opens issues with titles like 'Report: This is the Real Title' or 'Watch: This is the Real Title'. However most of the time the use of og:title and h1 allows extracting only 'This is the Real Title' -- and when this doesn't happen, the source is using the complete string as title also for the users (h1), so in some sense we can say that the full string is the intended output.

As a future development, the split function could be improved by excluding the part(s) containing the source name (e.g., via the domain name). However, still from empirical tests, these are now rare cases (but existing, e.g., hgtv.com). IMO this can be ignored (too rare cases for too much extra code).