Fixed logging messages and level + comment fix.
In the function set_reddit_top_img of article.py I edited the logging messages such that they made more sense, as well as not giving it such a high logging level. We were getting critical errors about "jpeg error with PIL, timed out". However, it seemed like it was simply images.py which was having a time out issue when downloading images for dimension measuring. I don't believe it had anything to do with PIL.