Adding MIME type to Article class
This change adds a mime_type field with associated setters to the
Article class. It also introduces changes to the network set of
functions, namely th 2xx download function with a DTO class that
ferries both the HTML and the MIME type form the requests data to the
Article.

In case of explicitly setting the input_html parameter in the
download() method in Article, the solution for now is to hardcode the
MIME type to "text/html".