Use urllib's ParseResult to check for valid URL hosts
The previous check was that:
- mailto: did not exist anywhere in the URL string
- http or https was found anywhere in the URL string

There are a couple places where this could (a) go wrong and (b)
be made faster.

Regarding the comment, "not sure if these rules are redundant,"
the answer is "they are and they aren't."  RFC 3986 only allows
one scheme to lead the URL.  *However*, this does not preclude
the possibility that another scheme, as a string, may be found
in the query string of the full canonical URL.

Using urllib.parse.ParseResult and checking its `scheme` attribute
against ALLOWED_SCHEMES kills two birds with one stone in this case.