Fix: re-allow alpha-character representation of month in DATE_REGEX
Following changes in 74763b6, this brings the updated
regex better inline with the original in allowing alpha
character abbreviations for month.

The previous DATE_REGEX in full (checkout e34c777 to see) was:

```
([\./\-_]{0,1}(19|20)\d{2})       # 4-digit year in 1900-2099
[\./\-_]{0,1}                     # optional sep
(
    ([0-3]{0,1}[0-9][\./\-_])     # 0-9 or 00 thru 39, followed by optional sep
    |                             # OR
    (\w{3,5}[\./\-_])             # 3-5 word characters [a-zA-Z0-9_], followed by optional sep
)
([0-3]{0,1}[0-9][\./\-]{0,1})?    # 0-9 or 00 thru 39, followed by optional sep
```

Note that this does *not* allow 2-digit year as was mentioned in the docstring
for valid_url(), so the updated version stays with the practice of only
allowing 4-digit years.  Additionally, the old regex allowed what would
have been impossible dates based on the calendar, such as 39 representing
the month or day.  These irregularities are removed in the new version
(as of 74763b6).

Note that we need the negative lookahead (?![a-zA-Z]) to ensure an incorrect
partial match is not made.