Use lazy string formatting in logging calls
Logging using traditional string formatting in logging calls, i.e.

    logging.warning('%s before you %s' % ('Look', 'leap!'))

is a fairly big performance hit since the entire expression will be
evaluated before the function call, rather than lazily, even if the
logging.warning call is not reached or is not emitted because of its
level and the levels of the corresponding handlers.

The logging package has a design specifically built to avoid this
and format strings lazily:

    logging.warning('%s before you %s', 'Look', 'leap!')

This second call syntax is used in some places, but not all,
in the newspaper package prior to this commit.

Internally the logging package uses %-formatting to merge the format
string and the variable arguments.  That is done via .makeRecord
in logging/__init__.py

Source:
https://docs.python.org/3.5/howto/logging.html#logging-variable-data

Found these calls via:

    grep -nr --include=\*.py "log\." .

Lastly, exchange log.warn() (deprecated) for log.warning().